seems to work somehow with booting the sensor - humidity changes, but temperature is very slow to change

+++ src/lib/ble/services.rs
@@ -11,9 +11,9 @@ pub struct BatteryService {
 #[nrf_softdevice::gatt_service(uuid = "181a")]
 pub struct EnviroSensingService {    
     #[characteristic(uuid = "2a6e", read, notify)]
-    pub temperature: u32,
+    pub temperature: i16,
     #[characteristic(uuid = "2a6f", read, notify)]
-    pub humidity: u32,
+    pub humidity: u16,
     #[characteristic(uuid = "2a6d", read, notify)]
     pub pressure: u32,
 }
